16.03 10-17

+++ WindowsFormsApplication1/Classes/MembershipFunctions/MembershipFunctionBase.cs
@@ -8,10 +8,10 @@ namespace WindowsFormsApplication1.Classes.MembershipFunctions
 {
     abstract public class MembershipFunctionBase
     {
-        public int a;
-        public int b;
-        public int c;
-        public int d;
+        public double a;
+        public double b;
+        public double c;
+        public double d;
 
         public abstract double CalculateFunctionValue(double _val_1/*, double _a, double _b = 0, double _c = 0, double _d = 0*/);
     }
